+++ CHANGELOG.md
@@ -17,6 +17,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- [log] Panic when parsing cookie headers
+
 ### Security
 
 ## [v1.91.0] - 2026-02-06